refactor the initial codebase to separate components
