+++ src/components/Home.jsx
@@ -35,21 +35,6 @@ const Home = () => {
     history.push('/login');
   };
 
-  const resourceServerExamples = [
-    {
-      label: 'Node/Express Resource Server Example',
-      url: 'https://github.com/okta/samples-nodejs-express-4/tree/master/resource-server',
-    },
-    {
-      label: 'Java/Spring MVC Resource Server Example',
-      url: 'https://github.com/okta/samples-java-spring/tree/master/resource-server',
-    },
-    {
-      label: 'ASP.NET Core Resource Server Example',
-      url: 'https://github.com/okta/samples-aspnetcore/tree/master/samples-aspnetcore-3x/resource-server',
-    },
-  ];
-
   if (authState.isPending) {
     return (
       <div>Loading...</div>
@@ -71,27 +56,6 @@ const Home = () => {
             Welcome, &nbsp;
             {userInfo.name}
             !
-          </p>
-          <p>
-            You have successfully authenticated against your Okta org, and have been redirected back to this application.  You now have an ID token and access token in local storage.
-            Visit the
-            {' '}
-            <a href="/profile">My Profile</a>
-            {' '}
-            page to take a look inside the ID token.
-          </p>
-          <h3>Next Steps</h3>
-          <p>Currently this application is a stand-alone front end application.  At this point you can use the access token to authenticate yourself against resource servers that you control.</p>
-          <p>This sample is designed to work with one of our resource server examples.  To see access token authentication in action, please download one of these resource server examples:</p>
-          <ul>
-            {resourceServerExamples.map((example) => <li key={example.url}><a href={example.url}>{example.label}</a></li>)}
-          </ul>
-          <p>
-            Once you have downloaded and started the example resource server, you can visit the
-            {' '}
-            <a href="/messages">My Messages</a>
-            {' '}
-            page to see the authentication process in action.
           </p>
         </div>
         )}
